feat: update docs and refactor Agent class
