+++ src/types.ts
@@ -5,6 +5,21 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { Connection, SfProject } from '@salesforce/core';
+import { FileProperties } from '@salesforce/source-deploy-retrieve';
+
+// ====================================================
+//               Agent Creation Types
+// ====================================================
+/**
+ * Options for creating instances of agents from an org.
+ */
+export type AgentOptions = {
+  connection: Connection;
+  project?: SfProject;
+  name: string;
+};
+
 /**
  * An agent job spec is a list of job titles and descriptions
  * to be performed by the agent.
@@ -195,3 +210,176 @@ export type DraftAgentTopicsResponse = {
   errorMessage?: string;
   topicDrafts: DraftAgentTopics;
 };
+
+// ====================================================
+//               Agent Testing Types
+// ====================================================
+export type TestStatus = 'NEW' | 'IN_PROGRESS' | 'COMPLETED' | 'ERROR' | 'TERMINATED';
+
+export type AgentTestStartResponse = {
+  runId: string;
+  status: TestStatus;
+};
+
+export type AgentTestStatusResponse = {
+  status: TestStatus;
+  startTime: string;
+  endTime?: string;
+  errorMessage?: string;
+};
+
+export type TestCaseResult = {
+  status: TestStatus;
+  startTime: string;
+  endTime?: string;
+  inputs: {
+    utterance: string;
+  };
+  generatedData: {
+    actionsSequence: string[];
+    outcome: string;
+    topic: string;
+  };
+  testResults: Array<{
+    name: string;
+    actualValue: string;
+    expectedValue: string;
+    score: number;
+    result: 'PASS' | 'FAILURE';
+    metricLabel: 'Accuracy' | 'Precision';
+    metricExplainability: string;
+    status: TestStatus;
+    startTime: string;
+    endTime?: string;
+    errorCode?: string;
+    errorMessage?: string;
+  }>;
+  testNumber: number;
+};
+
+export type AgentTestResultsResponse = {
+  status: TestStatus;
+  startTime: string;
+  endTime?: string;
+  errorMessage?: string;
+  subjectName: string;
+  testCases: TestCaseResult[];
+};
+
+export type AvailableDefinition = Omit<FileProperties, 'manageableState' | 'namespacePrefix'>;
+
+export type TestCase = {
+  utterance: string;
+  expectedActions: string[] | undefined;
+  expectedOutcome: string | undefined;
+  expectedTopic: string | undefined;
+};
+
+export type TestSpec = {
+  name: string;
+  description?: string;
+  subjectType: string;
+  subjectName: string;
+  subjectVersion?: string;
+  testCases: TestCase[];
+};
+
+export type AiEvaluationDefinition = {
+  AiEvaluationDefinition: {
+    description?: string;
+    name: string;
+    subjectType: 'AGENT';
+    subjectName: string;
+    subjectVersion?: string;
+    testCase: Array<{
+      expectation: Array<{
+        name: string;
+        expectedValue: string;
+      }>;
+      inputs: {
+        utterance: string;
+      };
+    }>;
+  };
+};
+
+// ====================================================
+//               Agent Preview Types
+// ====================================================
+export type ApiStatus = {
+  status: 'UP' | 'DOWN';
+};
+
+type Href = { href: string };
+
+export type AgentPreviewError = {
+  status: number;
+  path: string;
+  requestId: string;
+  error: string;
+  message: string;
+  timestamp: number;
+};
+
+export type AgentPreviewMessageLinks = {
+  self: Href | null;
+  messages: Href | null;
+  session: Href | null;
+  end: Href | null;
+};
+
+export type MessageType =
+  | 'Inform'
+  | 'TextChunk'
+  | 'ProgressIndicator'
+  | 'Inquire'
+  | 'Confirm'
+  | 'Failure'
+  | 'Escalate'
+  | 'SessionEnded'
+  | 'EndOfTurn'
+  | 'Error';
+
+export type AgentPreviewMessage = {
+  type: MessageType;
+  id: string;
+  feedbackId: string;
+  planId: string;
+  isContentSafe: boolean;
+  message: string;
+  result: {
+    type: string;
+    property: string;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    value: any;
+  };
+  citedReferences: {
+    type: string;
+    value: string;
+  };
+};
+
+export type AgentPreviewStartResponse = {
+  sessionId: string;
+  _links: AgentPreviewMessageLinks;
+  messages: AgentPreviewMessage[];
+};
+
+export type AgentPreviewSendResponse = {
+  messages: AgentPreviewMessage[];
+  _links: AgentPreviewMessageLinks;
+};
+
+export type AgentPreviewEndMessage = {
+  type: string;
+  id: string;
+  reason: string;
+  feedbackId: string;
+};
+
+export type AgentPreviewEndResponse = {
+  messages: AgentPreviewEndMessage[];
+  _links: AgentPreviewMessageLinks;
+};
+
+export type EndReason = 'UserRequest' | 'Transfer' | 'Expiration' | 'Error' | 'Other';
